+++ CHANGES.rst
@@ -5,7 +5,8 @@ Changelog of lizard-task
 0.10 (unreleased)
 -----------------
 
-- Nothing changed yet.
+- Added task_logging decorator eliminating custom logging code from
+  task definition.
 
 
 0.9 (2012-05-16)